requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==3.2.16
django-bootstrap4==22.3
importlib-metadata==2.1.3
Pillow==9.3.0
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.11.0
